+++ Dosier_LED_Ver1_3_Universalplatinen_Controller/Functions.ino
@@ -31,7 +31,7 @@ String printDouble(double value, byte nachkomma)
 
 int get_key(unsigned int input)
 {
-  byte k;
+  int k;
   for (k = 0; k < NUM_KEYS; k++)
   {
     if (input < adc_key_val[k])
